services/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulsefm-backend"
version = "0.1.0"
description = "PulseFM backend API service."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.104.1",
  "uvicorn[standard]==0.24.0",
  "pydantic-settings",
  "google-cloud-tasks",
  "google-cloud-firestore",
  "google-cloud-storage",
  "modal",
  "pulsefm-models",
  "pulsefm-gcs",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pulsefm_backend*"]
